149486: IdSelection is now and has no named sublcasses anymore
